pyproject.toml
[project]
name = "FastCampus-MCP-A2A"
version = "0.1.0"
description = "FastCampus Online Lecture - MCP & A2A - Practice1: LangGraph & MCP & A2A"
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE" }
authors = [
    { name = "Hyunjun Jeon", email = "jeonhj920@gmail.com" }
]
dependencies = [
    "a2a-sdk>=0.3.0",
    "fastmcp>=2.11.0",
    "google-adk>=1.9.0",
    "langchain-google-genai>=2.1.8",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.28",
    "langgraph>=0.6.2",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "tavily-python>=0.3.8",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=14.1",
    "redis>=5.2.0",
    "aiohttp>=3.11.15",
    "arxiv>=2.2.0",
    "langchain>=0.3.27",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "httpx-sse>=0.4.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "ruff>=0.12.7",
    "ty>=0.0.1a16",
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning:langgraph.*",
    "ignore::UserWarning:pytest.*",
]
